Fix tests and generalise validators
